pyproject.toml
[project]
name = "mmidas-jax"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.1.0",
    "pytest>=8.3.2",
    "expecttest>=0.2.1",
    "wheel>=0.44.0",
    "jax-metal>=0.1.0",
    "pyrsistent>=0.20.0",
    "ipykernel>=6.29.5",
    "pip>=24.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
